replace v1 telemetry with v2, delete some unused files/components
