+++ src/components/index.ts
@@ -80,7 +80,6 @@ export { FeatureRoadMap } from './Own'
 
 // Cody
 export { CodyFeatureCard } from './CodyFeatureCard'
-export { CodySteps } from './cody/CodySteps'
 export { CodebaseAwareIntelligence } from './cody/CodebaseAwareIntelligence'
 export { AiGeneratedCode } from './cody/AiGeneratedCode'
 export { CodyVideoTab } from './cody/CodyVideoTab'
